+++ adapters/BidMachine/BidMachineAdapter/BannerAdLoader.swift
@@ -52,7 +52,10 @@ final class BannerAdLoader {
 
   private func loadWaterfallAd() {
     do {
-      try client.loadWaterfallBannerAd(size: adConfiguration.adSize, delegate: self) {
+      try client.loadWaterfallBannerAd(
+        size: adConfiguration.adSize, delegate: self,
+        placementId: Util.placementId(from: adConfiguration)
+      ) {
         [weak self] error in
         guard let self else { return }
         guard error == nil else {
@@ -80,7 +83,8 @@ final class BannerAdLoader {
 
     do {
       try client.loadRTBBannerAd(
-        with: bidResponse, size: adConfiguration.adSize, delegate: self, watermark: watermark
+        with: bidResponse, size: adConfiguration.adSize, delegate: self,
+        placementId: Util.placementId(from: adConfiguration), watermark: watermark
       ) {
         [weak self] error in
         guard let self else { return }