+++ adapters/BidMachine/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Next Version
 - Update the minimum iOS version to relax Google Mobile Ads SDK dependency to allow minor versions.
+- Updated to pass placement ID to BidMachine SDK when collecting signals and
+  loading ads.
 
 #### [Version 3.7.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/bidmachine/BidMachineAdapter-3.7.1.0.zip)
 - Verified compatibility with BidMachine SDK version 3.7.1.